backend/requirements/base.txt
# Core Django
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.5
python-decouple==3.8

# Database
psycopg2-binary==2.9.9
dj-database-url==2.1.0

# Caching & Message Queue
redis==5.0.1
django-redis==5.4.0
celery==5.3.4
kombu==5.3.4

# Authentication & Security
djangorestframework-simplejwt==5.3.1
django-ratelimit==4.1.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3

# Monitoring & Logging
celery[redis]==5.3.4
flower==2.0.1
